requirements.txt
# Python version: 3.8
face-recognition==1.2.2
flask==3.0.0
flask-cors
flask-wtf==1.2.1
python-dotenv
opencv-python